+++ src/components/pages/LandingPage/DragonBoi/DragonBoi.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import dragonBoi from '../../../assets/dragon-boi.svg';
+import dragonBoi from '../../../../assets/dragon-boi.svg';
 
 const DragonBoi = (): React.ReactElement => {
   return (
